src/third_party/blink/web_tests/sensor/orientation-sensor.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../http/tests/resources/sensor-helpers.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/services/device/public/mojom/sensor_provider.mojom.js"></script>
<script src="resources/generic-sensor-utils.js"></script>
<script src="resources/generic-sensor-tests.js"></script>
<script>

'use strict';

if (!window.testRunner)
    debug('This test cannot be run without the TestRunner');

const kDefaultReading = [
  [ 1, 0, 0, 0 ]  // 180 degrees around X axis.
];
const kRotationMatrix = [1,  0,  0,  0,
                         0, -1,  0,  0,
                         0,  0, -1,  0,
                         0,  0,  0,  1];

async function checkPopulateMatrix(sensorProvider, sensorType) {
  let sensorObject = new sensorType();

  // Throws with insufficient buffer space.
  assert_throws({ name: 'TypeError' }, () => sensorObject.populateMatrix(new Float32Array(15)));

  // Throws if no orientation data available.
  assert_throws({ name: 'NotReadableError' }, () => sensorObject.populateMatrix(new Float32Array(16)));

  if (window.SharedArrayBuffer) {
    // Throws if passed SharedArrayBuffer view.
    assert_throws({ name: 'TypeError' }, () => sensorObject.populateMatrix(new Float32Array(new SharedArrayBuffer(16))));
  }

  sensorObject.start();

  let mockSensor = await sensorProvider.getCreatedSensor(sensorType.name);
  await mockSensor.setSensorReading(kDefaultReading);
  await new Promise((resolve, reject) => {
    sensorObject.onreading = () => {
      try {
        // Works for all supported types.
        let rotationMatrix32 = new Float32Array(16);
        sensorObject.populateMatrix(rotationMatrix32);
        assert_array_equals(rotationMatrix32, kRotationMatrix);

        let rotationMatrix64 = new Float64Array(16);
        sensorObject.populateMatrix(rotationMatrix64);
        assert_array_equals(rotationMatrix64, kRotationMatrix);

        let rotationDOMMatrix = new DOMMatrix();
        sensorObject.populateMatrix(rotationDOMMatrix);
        assert_array_equals(rotationDOMMatrix.toFloat64Array(),
                            kRotationMatrix);

        // Sets every matrix element.
        rotationMatrix64.fill(123);
        sensorObject.populateMatrix(rotationMatrix64);
        assert_array_equals(rotationMatrix64, kRotationMatrix);

        sensorObject.stop();
        resolve(mockSensor);
      } catch (e) {
        reject(e);
      };
    }
    sensorObject.onerror = reject;
  });
  return mockSensor.removeConfigurationCalled();
}

const kReadings = {
  readings: kDefaultReading,
  expectedReadings: kDefaultReading,
  expectedRemappedReadings: [
    // For 'orientation.angle == 270', which is set for tests at
    // at SensorProxy::GetScreenOrientationAngle().
    [-0.707107, 0.707107, 0, 0]
  ]
}

runGenericSensorTests(
    AbsoluteOrientationSensor,
    kReadings,
    verifyQuatSensorReading,
    ['accelerometer', 'gyroscope', 'magnetometer']);

sensor_test((t, sensorProvider) => {
  return checkPopulateMatrix(
      sensorProvider,
      AbsoluteOrientationSensor);
}, 'Test AbsoluteOrientationSensor.populateMatrix() method works correctly.');

runGenericSensorTests(
    RelativeOrientationSensor,
    kReadings,
    verifyQuatSensorReading,
    ['accelerometer', 'gyroscope']);

sensor_test((t, sensorProvider) => {
  return checkPopulateMatrix(
      sensorProvider,
      RelativeOrientationSensor);
}, 'Test RelativeOrientationSensor.populateMatrix() method works correctly.');
</script>
